+++ .github/scripts/update-required-config-node-version.js
@@ -19,9 +19,9 @@
 const { join } = require("node:path");
 const { existsSync, readFileSync, writeFileSync } = require("node:fs");
 
-const filePath = "src/migration/config-node.ts";
+const filePath = "src/plugins/config-node-checker.ts";
 
-const versionRegex = /requiredVersion = \[([0-9], [0-9]+, [0-9]+)\]/;
+const versionRegex = /requiredVersion: \[number, number, number\] = \[([0-9], [0-9]+, [0-9]+)\]/;
 
 try {
 	console.log("Check if the required Config Node version is up to date...");