+++ resources/common.js
@@ -18,7 +18,7 @@ var FirebaseUI = FirebaseUI || (function () {
 	"use strict";
 
 	const i18n = function (key, tplStrs) {
-		return i18nFullOptions(key, "load-config", "validator", tplStrs);
+		return i18nFullOptions(key, "firebase-in", "validator", tplStrs);
 	};
 
 	const validators = {
@@ -130,9 +130,13 @@ var FirebaseUI = FirebaseUI || (function () {
 					return validateChild(value, opt);
 				}
 
+				// NR version
+				const match = /([0-9]{1,2})\.([0-9]{1,3})\.([0-9]{1,3})-?/.exec(RED.settings.version || "0.0.0");
+				match?.shift();
+
 				// If NR version >= 3.1.3 use the new validators
-				const redVersion = (RED.settings.version || "0.0.0").split(".").map((s) => Number(s));
-				if (redVersion[0] > 3 || (redVersion[0] === 3 && (redVersion[1] > 1 || (redVersion[1] === 1 && redVersion[2] >= 3))))
+				const [major, minor, patch] = match?.map((v) => parseInt(v, 10)) || [0, 0, 0];
+				if (major > 3 || (major === 3 && (minor > 1 || (minor === 1 && patch >= 3))))
 					return RED.validators.typedInput(options)(value, opt);
 
 				// Workaround for node-red#4440 to pass type